feat(minifier): add `pure` to side-effect free global constructor during DCE (#11270)
So that codegen can print /* #__PURE__# */ during codegen.

This mimics esbuild https://github.com/evanw/esbuild/commit/79187160bcf3ca0e24d0371cdaf0defcd2fc6824

closes ENG-48
